now we have 'All Episodes' and 'New Episodes'
