+++ app/src/androidTest/java/de/test/antennapod/ui/MainActivityTest.java
@@ -73,6 +73,12 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
 
         // all episodes
         openNavDrawer();
+        solo.clickOnText(solo.getString(R.string.all_episodes_label));
+        solo.waitForView(android.R.id.list);
+        assertEquals(solo.getString(R.string.all_episodes_label), getActionbarTitle());
+
+        // new episodes
+        openNavDrawer();
         solo.clickOnText(solo.getString(R.string.new_episodes_label));
         solo.waitForView(android.R.id.list);
         assertEquals(solo.getString(R.string.new_episodes_label), getActionbarTitle());
